fix(settings): clear empty model selections
